requirements.txt
altgraph==0.17.3
attrs==23.1.0
Babel==2.12.1
certifi==2023.7.22
cffi==1.15.1
cryptography==41.0.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
ldap3==2.9.1
numpy==1.26.0
openpyxl==3.1.2
outcome==1.2.0
pandas==2.1.0
pefile==2023.2.7
Pillow==10.0.1
pyasn1==0.5.0
pycparser==2.21
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pyodbc==4.0.39
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
selenium==4.18.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tabulate==0.9.0
tkcalendar==1.6.1
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.0.5
wsproto==1.2.0
XlsxWriter==3.1.6
